profile cover

+++ app/views/content/front/profile/album.blade.php
@@ -1,6 +1,6 @@
 @extends('containers.frontend')
 
-@section('title') {{ 'Профиль|фото' }} @stop
+@section('title') {{ 'Альбом '.$album->name }} @stop
 
 @section('main')
 
@@ -19,16 +19,22 @@
 							<a href="#null">
 								<img src="/{{ $image->thumb_big }}" alt="">
 							</a>
-							<a href="/image/delete/{{ $image->id }}" class="fa fa-times delete-image"></a>
+							@if(Auth::check() && Auth::user()->id==$image->user_id)
+								<a href="/image/delete/{{ $image->id }}" class="fa fa-times delete-image" onclick="return confirm('Удалить?')?true:false;"></a>
+							@endif
 							<div class="cont-box">
-								<div class="like">
-									Нравиться: 15
-								</div>
-								<div class="set">
-									<h4>Установить</h4>
-									<a href="/album/setalbumcover/{{ $album->id.'/'.$image->id }}">обложкой альбома</a>
-									<a href="#null">баннером профиля</a>
-								</div>
+								@if(Auth::check() && Auth::user()->id==$image->user_id)
+									<div class="like" imageid="{{ $image->id }}">
+										Нравиться: <span class="likes-count">{{ $image->likes?$image->likes:'0' }}</span>
+									</div>
+								@endif
+								@if(Auth::check() && Auth::user()->id==$image->user_id)
+									<div class="set">
+										<h4>Установить</h4>
+										<a href="/album/setalbumcover/{{ $album->id.'/'.$image->id }}">обложкой альбома</a>
+										<a href="/album/setprofilecover/{{ $image->id }}">баннером профиля</a>
+									</div>
+								@endif
 								<div class="share">
 									<h4>Поделиться</h4>
 									<a href="#null" class="facebook"></a>
@@ -48,26 +54,30 @@
 						<img src="{{ Common_helper::getUserAvatar($user->id) }}" alt="">
 					</a>
 					<div class="name">
-						<a href="#null">{{ $userinfo->name.' '.$userinfo->surname }}</a>
+						<a href="#null">{{ $userInfo->name.' '.$userInfo->surname }}</a>
 						<span class="online"></span>
 						<span class="status">PRO</span>
 					</div>
-					<span class="place">{{ $userinfo->city }}</span>
+					<span class="place">{{ $userInfo->city }}</span>
 					<div class="rait">Рейтинг:&nbsp;&nbsp;452.5</div>
-				</div>
+				</div>				
 				<div class="user-btns">
 					<a href="/{{ $user->alias }}/photo" class="btn-gray">Обратно в профиль</a>
-					<a href="#null" class="btn-purple">Написать сообщение</a>
-				</div>
-				<div class="album-btns">
-					<a href="#upload-photo" class="fancybox add_photo">Добавить Фото</a>
-					<a href="/album/delete/{{ $album->id }}" class="del_album" onclick="return confirm('Удалить альбом?')?true:false;">Удалить Альбом</a>
-				</div>
+					@if(Auth::check() && Auth::user()->id!=$user->id)
+						<a href="#null" class="btn-purple">Написать сообщение</a>
+					@endif
+				</div>				
+				@if(Auth::check() && Auth::user()->id==$user->id)
+					<div class="album-btns">
+						<a href="#upload-photo" class="fancybox add_photo">Добавить Фото</a>
+						<a href="/album/delete/{{ $album->id }}" class="del_album" onclick="return confirm('Удалить альбом?')?true:false;">Удалить Альбом</a>
+					</div>
+				@endif
 			</div>
 		</div>
 		<div class="custom-modal" id="upload-photo">
 			<div class="title">Загрузить фотографии</div>
-			<form action="/album/uploadimage" class="dropzone">
+			<form action="/image/uploadimage" class="dropzone">
 			  <div class="fallback">
 			    <input name="file" type="file" multiple />
 			  </div>
@@ -89,7 +99,7 @@
 	$(document).ready(function(){
 		Dropzone.autoDiscover = false;
 		var myDropzone = new Dropzone(".dropzone", { 
-			url: "/album/uploadimage/{{ $album->id }}",
+			url: "/image/uploadimage/{{ $album->id }}",
 			acceptedFiles: ".png, .jpg, .jpeg",
 			dictDefaultMessage: 'Перетащите картинки в эту область'
 		});
@@ -100,6 +110,25 @@
 		myDropzone.on('error',function(file,message){
 			$('.dz-upload').fadeOut('normal');
 			$.fancybox.update();	
+		})
+
+		$('.like').on('click',function(){
+			elem = $(this).find('.likes-count');
+			$.ajax({
+				url: '/image/like',
+				type: 'post',
+				datatype: 'json',
+				data: {
+					imageid: $(this).attr('imageid')
+				},
+				success: function(ret){
+					if(ret=='success'){
+						elem.text(parseInt(elem.text())+1);
+					} else {
+						alert('Вы уже отметили эту фотографию');
+					}
+				}
+			})
 		})
 	})
